+++ backend/src/main/java/com/ll/hotel/domain/review/review/dto/response/GetReviewResponse.java
@@ -1,4 +1,6 @@
-package com.ll.hotel.domain.review.review.dto;
+package com.ll.hotel.domain.review.review.dto.response;
+
+import com.ll.hotel.domain.review.review.dto.ReviewDto;
 
 import java.util.List;
