+++ backend/src/main/java/com/ll/hotel/domain/review/review/dto/response/HotelReviewListResponse.java
@@ -1,0 +1,9 @@
+package com.ll.hotel.domain.review.review.dto.response;
+
+import com.ll.hotel.standard.page.dto.PageDto;
+
+public record HotelReviewListResponse (
+        PageDto<HotelReviewResponse> hotelReviewPage,
+        Double averageRating
+        ){
+}
