+++ backend/src/main/java/com/ll/hotel/domain/review/review/dto/response/MyReviewWithCommentDto.java
@@ -1,21 +1,22 @@
-package com.ll.hotel.domain.review.review.dto;
+package com.ll.hotel.domain.review.review.dto.response;
 
 import com.ll.hotel.domain.review.comment.dto.ReviewCommentDto;
 import com.ll.hotel.domain.review.comment.entity.ReviewComment;
+import com.ll.hotel.domain.review.review.dto.ReviewDto;
 import com.ll.hotel.domain.review.review.entity.Review;
 
 import java.time.LocalDateTime;
 
-public record HotelReviewWithCommentDto(
-        String memberEmail,
+public record MyReviewWithCommentDto(
+        String hotelName,
         String roomTypeName,
         ReviewDto reviewDto,
         ReviewCommentDto reviewCommentDto,
         LocalDateTime reservedAt
 ) {
-    public HotelReviewWithCommentDto(String memberEmail, String roomTypeName, Review review, ReviewComment reviewComment, LocalDateTime reservedAt) {
+    public MyReviewWithCommentDto(String hotelName, String roomTypeName, Review review, ReviewComment reviewComment, LocalDateTime reservedAt) {
         this(
-                memberEmail,
+                hotelName,
                 roomTypeName,
                 new ReviewDto(review),
                 reviewComment != null ? new ReviewCommentDto(reviewComment) : null,
